pyproject.toml
[project]
name = "effective-mobile-task"
version = "0.1.0"
description = ""
authors = [
    {name = "Ivan",email = "vasilev.i.b@edu.mirea.ru"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "flask (>=3.1.3,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "marshmallow (>=4.2.3,<5.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "dependency-injector (>=4.49.0,<5.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "scripts.dev:main"